+++ src/main/java/io/jenkins/infra/repository_permissions_updater/ArtifactoryAPI.java
@@ -10,8 +10,6 @@ import java.util.List;
 import java.util.logging.Level;
 
 public sealed interface ArtifactoryAPI permits ArtifactoryImpl {
-    /* Singleton support */
-
     /**
      * List all permission targets whose name starts with the configured prefix.
      *
